add ability to prepend content to an existing page
currently this is accomplished in an inefficient manner. The GET request for page content returns HTML, while we are posting in WIKI markup. I didn't want to pull in a separate library to convert from HTML to WIKI (or finding a conversion from markdown to confluence HTML).
Instead I retrieve the original content in HTML, publish the WIKI markup (replacing all existing content), then retreive the new HTML again. After that I combine the two and republish the consolidated HTML content.